+++ packages/mrm-core/src/formats/yaml.js
@@ -1,8 +1,13 @@
 // @ts-check
 const _ = require('lodash');
-const yaml = require('js-yaml');
+const yaml = require('yaml');
 const merge = require('../util/merge');
 const base = require('./file');
+
+const YAML_READ_OPTIONS = {
+	prettyErrors: true,
+};
+const DEFAULT_INDENT = 2;
 
 /**
  * @param {string} filename
@@ -11,7 +16,7 @@ const base = require('./file');
 module.exports = function(filename, defaultValues) {
 	const file = base(filename);
 
-	let json = yaml.safeLoad(file.get()) || defaultValues || {};
+	let json = yaml.parse(file.get(), YAML_READ_OPTIONS) || defaultValues || {};
 
 	return {
 		/** Return true if a file exists */
@@ -66,8 +71,8 @@ module.exports = function(filename, defaultValues) {
 
 		/** Save file */
 		save() {
-			const content = yaml.safeDump(json, {
-				lineWidth: 120,
+			const content = yaml.stringify(json, {
+				indent: file.getStyle().indent_size || DEFAULT_INDENT,
 			});
 			file.save(content);
 			return this;