+++ packages/mrm-task-readme/index.spec.js
@@ -4,10 +4,10 @@ jest.mock('mrm-core/src/util/log', () => ({
 	added: jest.fn(),
 }));
 
-const fs = require.requireActual('fs');
+const fs = jest.requireActual('fs');
 const path = require('path');
 const { omitBy } = require('lodash');
-const { getConfigGetter } = require('mrm');
+const { getTaskOptions } = require('mrm');
 const vol = require('memfs').vol;
 const task = require('./index');
 
@@ -15,21 +15,21 @@ const stringify = o => JSON.stringify(o, null, '  ');
 
 afterEach(() => vol.reset());
 
-it('should add a readme', () => {
+it('should add a readme', async () => {
 	vol.fromJSON({
 		[`${__dirname}/templates/Readme.md`]: fs
 			.readFileSync(path.join(__dirname, 'templates/Readme.md'))
 			.toString(),
 		'/package.json': stringify({
 			name: 'unicorn',
+			repository: 'gandalf/unicorn',
 		}),
 	});
 
 	task(
-		getConfigGetter({
-			name: 'Gendalf',
+		await getTaskOptions(task, false, {
+			name: 'Gandalf',
 			url: 'https://middleearth.com',
-			github: 'gendalf',
 		})
 	);