+++ packages/mrm-task-semantic-release/index.js
@@ -1,28 +1,47 @@
-const fs = require('fs');
-const gitUsername = require('git-username');
-const {
-	MrmError,
-	packageJson,
-	json,
-	yaml,
-	markdown,
-	uninstall,
-} = require('mrm-core');
+const packageRepoUrl = require('package-repo-url');
+const { packageJson, yaml, markdown, uninstall } = require('mrm-core');
 
-module.exports = function task({
-	github,
-	readmeFile,
-	semanticConfig,
-	semanticArgs,
-	semanticPeerDependencies,
-	semanticPreset,
-}) {
-	// Require .travis.yml
-	if (!fs.existsSync('.travis.yml')) {
-		throw new MrmError(
-			`Run travis task first:
-  mrm travis`
-		);
+module.exports = function task({ workflowFile, readmeFile }) {
+	// Create workflow file (no update)
+	const workflowYml = yaml(workflowFile);
+	if (!workflowYml.exists()) {
+		workflowYml
+			.merge({
+				name: 'Semantic Release',
+				on: {
+					push: {
+						branches: ['master'],
+					},
+				},
+				jobs: {
+					release: {
+						name: 'Release',
+						'runs-on': 'ubuntu-latest',
+						steps: [
+							{
+								uses: 'actions/checkout@v2',
+							},
+							{
+								uses: 'actions/setup-node@v1',
+								with: {
+									'node-version': 12,
+								},
+							},
+							{
+								run: 'npm ci',
+							},
+							{
+								env: {
+									GITHUB_TOKEN: '${{ secrets.GITHUB_TOKEN }}',
+									NPM_TOKEN: '${{ secrets.NPM_TOKEN }}',
+								},
+								run: 'npx semantic-release',
+							},
+						],
+					},
+				},
+			})
+			.save();
 	}
 
 	// package.json
@@ -32,71 +51,26 @@ module.exports = function task({
 	pkg.removeScript('semantic-release');
 	pkg.removeScript('travis-deploy-once');
 
-	// Add or remove custom semantic-release config
-	if (semanticConfig) {
-		pkg.merge({
-			release: semanticConfig,
-		});
-	} else {
-		pkg.unset('release');
-	}
+	// Remove custom semantic-release config
+	pkg.unset('release');
 
 	// Save package.json
 	pkg.save();
 
+	// Remove semantic-release from project dependencies
+	uninstall(['semantic-release', 'travis-deploy-once']);
+
+	// Remove semantic-release runner from commands list on .travis.yml
 	const travisYml = yaml('.travis.yml');
-	const afterSuccess = travisYml.get('after_success');
-
-	// Remove the official semantic-release runner from commands list on .travis.yml
-	if (Array.isArray(afterSuccess)) {
-		travisYml.set(
-			'after_success',
-			afterSuccess.filter(cmd => cmd !== 'npm run semantic-release')
-		);
-	}
-
-	// Add semantic-release runner to .travis.yml
-	const dependencies = [
-		...semanticPeerDependencies,
-		semanticPreset ? `conventional-changelog-${semanticPreset}` : null,
-	].filter(Boolean);
-	const commands = [
-		dependencies.length > 0
-			? `npm install --no-save ${dependencies.join(' ')}`
-			: null,
-		`npx semantic-release ${semanticArgs}`.trim(),
-	].filter(Boolean);
-	travisYml
-		.merge({
-			after_success: commands,
-			branches: {
-				except: ['/^v\\d+\\.\\d+\\.\\d+$/'],
-			},
-		})
-		.save();
-
-	// Add semantic-release config if we’re using a preset
-	if (semanticPreset) {
-		json('.releaserc.json')
-			.set({
-				plugins: [
-					[
-						'@semantic-release/commit-analyzer',
-						{
-							preset: semanticPreset,
-						},
-					],
-					[
-						'@semantic-release/release-notes-generator',
-						{
-							preset: semanticPreset,
-						},
-					],
-					'@semantic-release/npm',
-					'@semantic-release/github',
-				],
-			})
-			.save();
+	if (travisYml.exists()) {
+		const afterSuccess = travisYml.get('after_success');
+		if (Array.isArray(afterSuccess)) {
+			travisYml.set(
+				'after_success',
+				afterSuccess.filter(cmd => !cmd.includes('semantic-release'))
+			);
+		}
+		travisYml.save();
 	}
 
 	// Add npm package version badge to Readme.md
@@ -112,11 +86,8 @@ module.exports = function task({
 			.save();
 	}
 
-	// Remove semantic-release from project dependencies
-	uninstall(['semantic-release', 'travis-deploy-once']);
-
 	console.log(
-		`Please setup semantic-release authentication on Travis CI:
+		`Please add secrets on GitHub:
 
 1. Generate npm token:
 
@@ -134,48 +105,28 @@ module.exports = function task({
 
    More info: https://help.github.com/en/articles/creating-a-personal-access-token-for-the-command-line
 
-3. Add environment variables to Travis CI:
+3. Add environment variables to GitHub repository:
 
-   https://travis-ci.org/${github}/${packageName}/settings
+   ${packageRepoUrl()}/settings/secrets
 
    NPM_TOKEN = <Generated npm token>
    GH_TOKEN = <Generated GitHub token>
 
-   DO NOT enable “Display value in build log”.
-
-   More info: https://semantic-release.gitbook.io/semantic-release/usage/ci-configuration
+   More info: https://github.com/semantic-release/semantic-release/blob/beta/docs/recipes/github-actions.md
 `
 	);
 };
 
 module.exports.description = 'Adds semantic-release';
 module.exports.parameters = {
-	github: {
+	workflowFile: {
 		type: 'input',
-		message: 'Enter your GitHub username',
-		default: gitUsername(),
-		validate(value) {
-			return value ? true : 'GitHub username is required';
-		},
+		message: 'Enter location of GitHub Actions workflow file',
+		default: '.github/workflows/release.yml',
 	},
 	readmeFile: {
 		type: 'input',
 		message: 'Enter filename for readme',
 		default: 'Readme.md',
-	},
-	semanticArgs: {
-		type: 'input',
-		message: 'Enter arguments for semantic-release command line tool',
-		default: '',
-	},
-	semanticConfig: {
-		type: 'config',
-	},
-	semanticPeerDependencies: {
-		type: 'config',
-		default: [],
-	},
-	semanticPreset: {
-		type: 'config',
 	},
 };