+++ packages/mrm-task-contributing/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.12](https://github.com/sapegin/mrm/compare/mrm-task-contributing@2.0.11...mrm-task-contributing@2.0.12) (2020-10-27)
+
+**Note:** Version bump only for package mrm-task-contributing
+
+
+
+
+
+## [2.0.11](https://github.com/sapegin/mrm/compare/mrm-task-contributing@2.0.10...mrm-task-contributing@2.0.11) (2020-10-27)
+
+**Note:** Version bump only for package mrm-task-contributing
+
+
+
+
+
+## [2.0.10](https://github.com/sapegin/mrm/compare/mrm-task-contributing@2.0.9...mrm-task-contributing@2.0.10) (2020-10-27)
+
+**Note:** Version bump only for package mrm-task-contributing
+
+
+
+
+
 ## [2.0.9](https://github.com/sapegin/mrm/compare/mrm-task-contributing@2.0.8...mrm-task-contributing@2.0.9) (2020-10-26)
 
 **Note:** Version bump only for package mrm-task-contributing